% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspre.R
\name{approx_ImGi}
\alias{approx_ImGi}
\title{Approximates(I-G)^-1 by truncated sum I + G + G%^%2 + ...}
\usage{
approx_ImGi(G, k, keep = TRUE)
}
\arguments{
\item{G}{D x D graph}

\item{k}{Integer. Number of powers to use to approximate the inverse}

\item{keep}{Bool. Default TRUE to keep each power for downstream computation}
}
\description{
Approximates(I-G)^-1 by truncated sum I + G + G%^%2 + ...
}
