% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspre_h5.R
\name{calc_inst_effect_h5X}
\alias{calc_inst_effect_h5X}
\title{Calculates correlation and standard error of instrument with target feature.}
\usage{
calc_inst_effect_h5X(inst_id, target, X, X_control, X_ids, X_vars)
}
\arguments{
\item{inst_id}{String. Id of instrumental to calculate outcomes from.}

\item{target}{String. Target feature of the instrument.}

\item{X}{H5D. Example: `X=hfile[['X']]` where hfile is an hdf5r object
containing a data matrix stored under label 'X'.}

\item{X_control}{Matrix. features by control observations. Usually
pulled from X but stored in memory so you don't have to pull the
same control observations off of disk for every instrument.}

\item{X_ids}{Sequence of strings with length equal to number of columns in
X. Entries correspond to the instrument applied to that entry.}

\item{X_vars}{Sequence of strings with length equal to the number of rows
in X. Name of the feature measured in that row.}
}
\description{
Calculates correlation and standard error of instrument with target feature.
}
