% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{cor_w_se}
\alias{cor_w_se}
\title{Estimates correlation matrix of X in the presence of missing data.}
\usage{
cor_w_se(X)
}
\arguments{
\item{X}{NxD matrix with optionally missing entries.}
}
\description{
Uses simple approximation sqrt((1-r^2)/(n-2)) which is technically
incorrect but works well in practice.
}
