% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{dataset_to_score}
\alias{dataset_to_score}
\title{Converts simulated dataset to score object for use with GIES function.}
\usage{
dataset_to_score(dataset)
}
\arguments{
\item{dataset}{Dataset generated by `generate_dataset`}
}
\description{
Converts simulated dataset to score object for use with GIES function.
}
