% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspre.R
\name{est_exp_var}
\alias{est_exp_var}
\title{Estimates total variance explained in and by each gene.}
\usage{
est_exp_var(H, s2, v)
}
\arguments{
\item{H}{DxD matrix.}

\item{s2}{D-vector. Noise variance of each node.}

\item{v}{D-vector. Total variance of each node.}
}
\description{
Estimates total variance explained in and by each gene.
}
