% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspre.R
\name{fit_inspre_sequence}
\alias{fit_inspre_sequence}
\title{Fits inspre model for sequence of lambda values.}
\usage{
fit_inspre_sequence(
  X,
  lambda,
  W = NULL,
  rho = 1,
  its = 100,
  delta_target = 1e-04,
  symmetrize = FALSE,
  verbose = 1,
  gamma = 0,
  train_prop = 1,
  mu = 5,
  tau = 2.5,
  solve_its = 3,
  ncores = 1,
  warm_start = FALSE
)
}
\arguments{
\item{X}{DxD Matrix to find approximate sparse inverse of.}

\item{lambda}{Float or sequence of floats. Path of L1 regularization strength
on inverse of X.}

\item{W}{DxD Matrix of weights.}

\item{rho}{Float. Initial learning rate for ADMM.}

\item{its}{Integer. Maximum number of iterations.}

\item{delta_target}{Float. Target change in solution.}

\item{symmetrize}{True to force the output to be symmetric. If the input
is symmetric, the output isn't always perfectly symmetric due to numerical
issues.}

\item{verbose}{0, 1 or 2. 2 to print convergence progress for each lambda,
1 to print convergence result for each lambda, 0 for no output.}

\item{gamma}{Float or sequence of floats. Determinant penalty strength.}

\item{train_prop}{Proportion of the dataset to use to train the model. Set
to 1.0 to use all data.}

\item{mu}{rho modification parameter for ADMM. Rho will be
increased/decreased when the dual constrant and primal constraint are off
by a factor of > mu.}

\item{tau}{rho modification parameter for ADMM. When called for, rho will be
increased/decreased by the factor tau.}

\item{solve_its}{Integer, number of iterations of bicgstab/lasso to run
for each U and V update.}

\item{ncores}{Integer, number of cores to use.}

\item{warm_start}{Boolean, TRUE to start next fit with result of previous
fit. Default FALSE.}
}
\description{
Fits inspre model for sequence of lambda values.
}
