% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{fit_R_MR}
\alias{fit_R_MR}
\title{Uses MR to fit R matrix}
\usage{
fit_R_MR(input_data, mr_method, mr_function = NULL)
}
\arguments{
\item{input_data}{A dataframe with columns `Exposure`, `Outcome`, `inst`,
`beta_exp`, `beta_out`, `se_exp` and `se_out`. Each column is a list-column
with values for the corresponding instrument for that pair of phenotypes.}

\item{mr_method}{MR Method to use. Currently supported: inverse variance
weighting (IVW).}

\item{mr_function}{function taking 4 inputs `b_exp`, `b_out`, `se_exp`
and `se_out`. Use to specify a custom MR function not implemented above.}
}
\description{
Uses MR to fit R matrix
}
