% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_U_VU_const}
\alias{fit_U_VU_const}
\title{Fit U satisfying VU = I}
\usage{
fit_U_VU_const(
  WX,
  W,
  V,
  U,
  theta,
  rho,
  gamma = 0,
  solve_its = 10L,
  fixd = FALSE,
  nthreads = 1L
)
}
\description{
Fit U satisfying VU = I
}
