% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_V_VU_const}
\alias{fit_V_VU_const}
\title{Fit V satisfying VU = I}
\usage{
fit_V_VU_const(
  V,
  U,
  theta,
  rho,
  lambda = 0,
  solve_its = 10L,
  fixd = FALSE,
  nthreads = 1L
)
}
\description{
Fit V satisfying VU = I
}
