% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_cor}
\alias{generate_cor}
\title{Generates correlation/precision matrices for gaussian graphical model.}
\usage{
generate_cor(D, p_net, sd_net = 1)
}
\arguments{
\item{D}{Integer. Dimensionality.}

\item{p_net}{Float. Connection density.}

\item{sd_net}{Float. Standard deviation of initial covariance.}
}
\description{
Generates correlation/precision matrices for gaussian graphical model.
}
