% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_network}
\alias{generate_network}
\title{Simulates random graph, optionally a DAG}
\usage{
generate_network(
  D,
  graph = "scalefree",
  p = 0.4,
  v = 0.2,
  DAG = FALSE,
  max_eig = 0.9
)
}
\arguments{
\item{D}{Integer. Number of nodes to simulate.}

\item{graph}{String. One of 'scalefree' or 'random'. Type of network
to simulate.}

\item{p}{Float between 0 and 1. Parameter for network generation. For
scalefree, p is the probability of adding a new node at each iteration.
For random, it is the probability of including each edge.}

\item{v}{Float. Mode of the pert distribution for simulated edge weights.}

\item{DAG}{Bool. TRUE to ensure the returned graph is a DAG.}
}
\description{
Edges are sampled from a PERT distribution with random sign. By
default these are between 0.05 and 0.8 with a mode of 0.2 and
intended to represent "root variance-explained".
}
