% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{get_direct}
\alias{get_direct}
\title{Fits exact model to data.}
\usage{
get_direct(R)
}
\arguments{
\item{R}{D x D matrix of "total causal effects".}
}
\value{
D x D matrix with zero diagonal of deconvoluted direct effects.
}
\description{
Fits exact model to data.
}
