% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspre.R
\name{ilasso}
\alias{ilasso}
\title{Fits simpler inverse LASSO model.}
\usage{
ilasso(X, lambda, its = 100, rmse_target = 1e-04, verbose = FALSE)
}
\arguments{
\item{X}{DxD Matrix to find approximate sparse inverse of.}

\item{lambda}{Float. L1 regularization strength on inverse of X.}

\item{its}{Integer. Number of iterations.}

\item{rmse_target}{Float. Target change in solution.}

\item{verbose}{Bool. True to print solution progress.}
}
\description{
Fits simpler inverse LASSO model.
}
