% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspre.R
\name{ilasso_worker}
\alias{ilasso_worker}
\title{Fits simpler inverse LASSO model for one lambda value.}
\usage{
ilasso_worker(
  X,
  lambda,
  lambda_min_ratio = 0.001,
  its = 1000,
  delta_target = 1e-04,
  warm_start = NULL,
  verbose = FALSE
)
}
\arguments{
\item{X}{DxD Matrix to find approximate sparse inverse of.}

\item{lambda}{Float, sequence of floats of NULL. L1 regularization strength
on inverse of X. If NULL, a logarithmicallly spaced set of values between
the maximimum absolute off diagonal element of X and lambda_min_ratio
times this value will be used.}

\item{lambda_min_ratio}{Float, ratio of maximum lambda to minimum lambda.}

\item{its}{Integer. Maximum number of iterations.}

\item{delta_target}{Float. Target change in solution.}

\item{warm_start}{DxD matrix, initial solution.}

\item{verbose}{Bool. True to print solution progress.}
}
\description{
Fits simpler inverse LASSO model for one lambda value.
}
