% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspre.R
\name{inspre}
\alias{inspre}
\title{Finds the inverse of X using Inverse Sparse Regression.}
\usage{
inspre(
  X,
  W,
  rho = 0.1,
  lambda = 0.01,
  gamma = 0.01,
  its = 100,
  rmse_target = 1e-04,
  verbose = FALSE
)
}
\arguments{
\item{X}{DxD Matrix to find approximate sparse inverse of.}

\item{W}{DxD Matrix of weights.}

\item{rho}{Float, learning rate for ADMM.}

\item{lambda}{Float. L1 regularization strength on inverse of X.}

\item{gamma}{Float. L2 regularization on approximation to X that is inverted.}

\item{its}{Integer. Number of iterations.}

\item{rmse_target}{Float. Target change in solution.}

\item{verbose}{Bool. True to print solution progress.}
}
\description{
Finds the inverse of X using Inverse Sparse Regression.
}
