% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspre.R
\name{inspre_worker}
\alias{inspre_worker}
\title{Worker function to fit inspre for a single value of lambda.}
\usage{
inspre_worker(
  X,
  W = NULL,
  rho = 1,
  lambda = 0.01,
  its = 100,
  delta_target = 1e-04,
  warm_start = NULL,
  symmetrize = FALSE,
  verbose = FALSE,
  gamma = 0,
  mu = 5,
  tau = 1.5,
  solve_its = 3,
  ncores = 1,
  constraint = "UV",
  DAG = FALSE
)
}
\arguments{
\item{X}{DxD Matrix to find approximate sparse inverse of.}

\item{W}{DxD Matrix of weights.}

\item{rho}{Float, learning rate for ADMM.}

\item{lambda}{Float. L1 regularization strength on inverse of X.}

\item{its}{Integer. Maximum number of iterations.}

\item{delta_target}{Float. Target change in solution.}

\item{warm_start}{List with elements "U" and "V" containing starting point.}

\item{symmetrize}{True to force the output to be symmetric. If the input
is symmetric, the output isn't always perfectly symmetric due to numerical
issues.}

\item{verbose}{Bool. True to print solution progress.}

\item{gamma}{Float. Determinant penalty strength.}

\item{mu}{rho modification parameter for ADMM. Rho will be
increased/decreased when the dual constrant and primal constraint are off
by a factor of > mu.}

\item{tau}{rho modification parameter for ADMM. When called for, rho will be
increased/decreased by the factor tau.}

\item{solve_its}{Integer, number of iterations of bicgstab/lasso to run
for each U and V update.}

\item{ncores}{Integer, number of cores to use.}

\item{constraint}{One of "UV" or "VU". Constraint to use.}

\item{DAG}{Bool. True to resitrict solutions to approximate DAGs.}
}
\description{
Worker function to fit inspre for a single value of lambda.
}
