% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{join_bigbrain}
\alias{join_bigbrain}
\title{Join bigbrain datasets that have been estimated on different chunks}
\usage{
join_bigbrain(...)
}
\arguments{
\item{...}{datasets to join}
}
\description{
Join bigbrain datasets that have been estimated on different chunks
}
