% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lasso_one_iteration}
\alias{lasso_one_iteration}
\title{Fit lasso model (one iteration)}
\usage{
lasso_one_iteration(X, Y, B, lambda)
}
\arguments{
\item{X}{NxP matrix of covariates}

\item{Y}{N-vector of response}

\item{B}{P-vector of coefficents (to be updated)}

\item{lambda}{P-vector L1 regulation coeffs}
}
\description{
Minimize .5 || Y - XB ||^2_2 + gamma |B|_1
}
