% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspre.R
\name{make_weights}
\alias{make_weights}
\title{Helper function to make weights for inspre.}
\usage{
make_weights(SE, max_weight = NULL)
}
\arguments{
\item{SE}{DxD matrix of standard errors.}

\item{max_weight}{Highest possible weight, useful when some SEs are
near 0.}
}
\description{
Helper function to make weights for inspre.
}
