% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspre_h5.R
\name{parse_hdf5_df}
\alias{parse_hdf5_df}
\title{Simple parses for hdf5 metadata into a data.frame.}
\usage{
parse_hdf5_df(hfile, entry = "obs")
}
\arguments{
\item{hfile}{hdf5r object.}

\item{entry}{String. Name of the entry in the hfile to parse.}
}
\description{
Note that this was written with GPWS data in mind, thus
it can be used to parse the 'obs' and 'var' matrices
that correspond to the columns and rows of the data object
stored under 'X' in those files. YMMV for other hdf5 files.
}
