% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspre.R
\name{predict_inspre}
\alias{predict_inspre}
\title{Calculates model predictions given new data, with cross-validation errors.}
\usage{
predict_inspre(res, .X, .beta, .targets)
}
\arguments{
\item{res}{Inspre result. Result of running fit_inspre for a single or
sequence of hyperparameter values.}

\item{X}{observations x features data matrix. Colnames represent feature
names. Rownames also correspond to feature names indicating an intervention
was applied to that feature in that sample, with "control" indicating
no intervention.}

\item{beta}{Sequence of floats. Obsevered scale estimates of instrument
effects.}
}
\description{
This function uses the mean((X - XG - ZB)**2) error formulation.
}
