% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{read_bigbrain}
\alias{read_bigbrain}
\title{Reads bigbrain formatted summary statistics.}
\usage{
read_bigbrain(
  filename,
  beta_col,
  p_col,
  id_col = "variant_id",
  se_col = NULL,
  p_thresh = 5e-08,
  chr_col = "chr",
  pos_col = "pos",
  ref_col = "ref",
  alt_col = "alt",
  feature_col = "feature",
  delim = "\\t"
)
}
\arguments{
\item{filename}{}
}
\description{
Reads the bigbrain formatter summary statistics file. Reads one line at
a time and saves each as Exposure, Outcome, Instrument, Effect size. Also
saves instrument information.
}
