% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{stars_cv}
\alias{stars_cv}
\title{Calculates STARS cross-validation.}
\usage{
stars_cv(X, method, train_prop = 0.8, cv_folds = 10)
}
\arguments{
\item{X}{N x D data matrix.}

\item{method}{Function that returns a list with two elements. "theta", a
D x D x nlambda matrix and "lambda" a list of lambda values used.}

\item{train_prop}{Float, proportion of rows of X to use at each iteration.}

\item{cv_folds}{Integer, number of cross-validation folds.}
}
\description{
Calculates STARS cross-validation.
}
