% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspre.R
\name{fit_inspre_sequence}
\alias{fit_inspre_sequence}
\title{Fits inspre model for sequence of lambda values.}
\usage{
fit_inspre_sequence(
  X,
  lambda,
  W = NULL,
  rho = 1,
  its = 1000,
  delta_target = 1e-04,
  symmetrize = FALSE,
  verbose = 1,
  detpen = 0,
  train_prop = 1,
  warm_start = NULL
)
}
\arguments{
\item{X}{DxD Matrix to find approximate sparse inverse of.}

\item{lambda}{Float or sequence of floats. Path of L1 regularization strength
on inverse of X.}

\item{W}{DxD Matrix of weights.}

\item{rho}{Float. Initial learning rate for ADMM.}

\item{its}{Integer. Maximum number of iterations.}

\item{delta_target}{Float. Target change in solution.}

\item{symmetrize}{True to force the output to be symmetric. If the input
is symmetric, the output isn't always perfectly symmetric due to numerical
issues.}

\item{verbose}{0, 1 or 2. 2 to print convergence progress for each lambda,
1 to print convergence result for each lambda, 0 for no output.}

\item{detpen}{Determinant penalty strength.}

\item{train_prop}{Proportion of the dataset to use to train the model. Set
to 1.0 to use all data.}
}
\description{
Fits inspre model for sequence of lambda values.
}
