% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_dataset}
\alias{generate_dataset}
\title{Generates correlation/precision matrices for gaussian graphical model.}
\usage{
generate_dataset(
  D,
  N,
  p_net,
  sd_net = 1,
  normalize = TRUE,
  min_missing = 0,
  max_missing = 0
)
}
\arguments{
\item{D}{Integer. Dimensionality.}

\item{N}{Integer. Number of samples to simulate.}

\item{p_net}{Float. Connection density.}

\item{sd_net}{Float. Standard deviation of initial covariance.}

\item{normalize}{Bool. True to return data and parameters with mean 0,
var 1.}

\item{min_missing}{Float. Minimum missing percentage. For each dimension
a value will be drawn uniformly between this and max_missing to determine
the probability an observation will be removed in that feature.}

\item{max_missing}{Float. Maximum missing percentage. See min_missing.}
}
\description{
Generates correlation/precision matrices for gaussian graphical model.
}
