% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_dataset}
\alias{generate_dataset}
\title{Simulates an intervention dataset.}
\usage{
generate_dataset(
  D,
  N_cont,
  N_int,
  int_beta = -2,
  graph = "scalefree",
  v = 0.2,
  p = 0.4,
  DAG = FALSE,
  C = floor(0.1 * D),
  noise = "gaussian",
  model = "inspre"
)
}
\arguments{
\item{D}{Integer. Number of nodes to simulate.}

\item{N_cont}{Integer. Number of control samples to simulate.}

\item{N_int}{Integer. Mean of number of intervention samples to simulate per node.}

\item{graph}{String. One of 'scalefree' or 'random'. Type of network
to simulate.}

\item{DAG}{Bool. TRUE to ensure the returned graph is a DAG.}

\item{C}{Integer. Number of fully connected confounding nodes to simulate.}

\item{noise}{String. Noise model to simulate, currently just "gaussian".}

\item{model}{Data generating model. One of "inspre" or "dotears".}

\item{size}{Float. Size parameter for NB distribution for int samples.}

\item{int_r2}{Float. Mean variance in node explained by intervention.}

\item{int_dir}{string. 'positive', 'negative', or 'both'. Direction of
effect of intervention on node.}

\item{v_mode}{Float. Mode of the pert distribution.}

\item{v_min}{Float. Min of the pert distribution.}

\item{v_max}{Float. Max of the pert distribuion.}
}
\description{
Simulates a graph-intervention dataset with corresponding graph.
}
