% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_network}
\alias{generate_network}
\title{Simulates random graph, optionally a DAG}
\usage{
generate_network(
  D,
  graph = "scale-free",
  v_mode = 0.2,
  v_max = 0.8,
  v_min = 0.05,
  DAG = FALSE
)
}
\arguments{
\item{D}{Integer. Number of nodes to simulate.}

\item{graph}{String. One of 'scale-free' or 'random'. Type of network
to simulate.}

\item{v_mode}{Float. Mode of the pert distribution.}

\item{v_max}{Float. Max of the pert distribuion.}

\item{v_min}{Float. Min of the pert distribution.}

\item{DAG}{Bool. TRUE to ensure the returned graph is a DAG.}
}
\description{
Edges are sampled from a PERT distribution with random sign. By
default these are between 0.05 and 0.8 with a mode of 0.2 and
intended to represent "root variance-explained".
}
