% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspre.R
\name{inspre}
\alias{inspre}
\title{Finds the inverse of X using Inverse Sparse Regression.}
\usage{
inspre(
  X,
  W = NULL,
  rho = 1,
  lambda = NULL,
  lambda_min_ratio = 0.01,
  nlambda = 20,
  alpha = 0,
  gamma = NULL,
  its = 100,
  delta_target = 1e-04,
  symmetrize = FALSE,
  verbose = 1,
  train_prop = 0.8,
  cv_folds = 0,
  mu = 5,
  tau = 1.5,
  solve_its = 3,
  ncores = 1
)
}
\arguments{
\item{X}{DxD Matrix to find approximate sparse inverse of.}

\item{W}{DxD Matrix of weights.}

\item{rho}{Float. Initial learning rate for ADMM.}

\item{lambda}{Float, sequence of floats of NULL. L1 regularization strength
on inverse of X. If NULL, a logarithmicallly spaced set of values between
the maximimum absolute off diagonal element of X and lambda_min_ratio
times this value will be used.}

\item{lambda_min_ratio}{Float, ratio of maximum lambda to minimum lambda.}

\item{nlambda}{Integer. Number of lambda values to try.}

\item{alpha}{Float between 0 and 1 or NULL. If > 0, the model will be fit
once with gamma = 0 to find L0, then all subsequent fits will use
gamma = alpha * L0 / D. Set to NULL to provide gamma directly.}

\item{gamma}{Float or sequence of nlambda floats or NULL. Determinant
regularization strength to use (for each lambda value). It is recommended
to set alpha rather than setting this directly.}

\item{its}{Integer. Maximum number of iterations.}

\item{delta_target}{Float. Target change in solution.}

\item{symmetrize}{True to force the output to be symmetric. If the input
is symmetric, the output isn't always perfectly symmetric due to numerical
issues.}

\item{verbose}{0, 1 or 2. 2 to print convergence progress for each lambda,
1 to print convergence result for each lambda, 0 for no output.}

\item{train_prop}{Float between 0 and 1. Proportion of data to use for
training in cross-validation.}

\item{cv_folds}{Integer. Number of cross-validation folds to perform.}

\item{mu}{rho modification parameter for ADMM. Rho will be
increased/decreased when the dual constrant and primal constraint are off
by a factor of > mu.}

\item{tau}{rho modification parameter for ADMM. When called for, rho will be
increased/decreased by the factor tau.}

\item{solve_its}{Integer, number of iterations of bicgstab/lasso to run
for each U and V update.}

\item{ncores}{Integer, number of cores to use.}
}
\description{
Finds the inverse of X using Inverse Sparse Regression.
}
