% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspre.R
\name{inspre_worker}
\alias{inspre_worker}
\title{Worker function to fit inspre for a single value of lambda.}
\usage{
inspre_worker(
  X,
  W = NULL,
  rho = 0.1,
  lambda = 0.01,
  gamma = 0,
  its = 1000,
  delta_target = 1e-04,
  warm_start = NULL,
  symmetrize = FALSE,
  verbose = FALSE,
  detpen = 0.1
)
}
\arguments{
\item{X}{DxD Matrix to find approximate sparse inverse of.}

\item{W}{DxD Matrix of weights.}

\item{rho}{Float, learning rate for ADMM.}

\item{lambda}{Float. L1 regularization strength on inverse of X.}

\item{gamma}{Float. L2 regularization on approximation to X that is inverted.}

\item{its}{Integer. Maximum number of iterations.}

\item{delta_target}{Float. Target change in solution.}

\item{warm_start}{List with elements "U" and "V" containing starting point.}

\item{symmetrize}{True to force the output to be symmetric. If the input
is symmetric, the output isn't always perfectly symmetric due to numerical
issues.}

\item{verbose}{Bool. True to print solution progress.}

\item{detpen}{Determinant penalty strength.}
}
\description{
TODO(brielin): Remove gamma penalty.
TODO(brielin): Infer detpen automatically.
}
