% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspre.R
\name{make_weights}
\alias{make_weights}
\title{Helper function to make weights for inspre.}
\usage{
make_weights(SE, max_min_ratio = NULL)
}
\arguments{
\item{SE}{DxD matrix of standard errors.}

\item{max_min_ratio}{Float > 1. Ratio of maximum weight to minimum non-zero
weight. Improves conditioning when some SEs are very small.}
}
\description{
Helper function to make weights for inspre.
}
