% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{read_bigbrain}
\alias{read_bigbrain}
\title{Reads bigbrain formatted summary statistics.}
\usage{
read_bigbrain(
  filename,
  beta_col,
  p_col,
  id_col = "variant_id",
  se_col = NULL,
  p_thresh = 5e-08,
  chr_col = "chr",
  pos_col = "pos",
  ref_col = "ref",
  alt_col = "alt",
  feature_col = "feature",
  delim = "\\t",
  filter_col = NULL,
  filter_max = NULL,
  filter_min = NULL
)
}
\arguments{
\item{filename}{String. Path to file you want to parse.}

\item{beta_col}{String. Name of the column containing beta.}

\item{p_col}{String. Name of the column containing p values.}

\item{id_col}{String. Name of the column containing variant ids (usually
rsids).}

\item{se_col}{String. Name of the column containing standard errors of beta
estimates.}

\item{p_thresh}{Float. Threshold to use for instrument inclusion.}

\item{chr_col}{String. Name of the column containing the SNP chromosome.}

\item{pos_col}{String. Name of the column containing the SNP position.}

\item{ref_col}{String. Name of the column containing the reference allele.}

\item{alt_col}{String. Name of the column containing the alternative allele.}

\item{feature_col}{String. Name of the column containing the feature.}

\item{delim}{String. File field delimiter.}

\item{filter_col}{String. Column to filter on. NULL for no additional filtering.}

\item{filter_max}{Float. Maximum value of `filter_col` for inclusion.}

\item{filter_min}{Float. Minimum value of `filter_col` for inclusion.}
}
\description{
Reads the bigbrain formatter summary statistics file. Reads one line at
a time and saves each as Exposure, Outcome, Instrument, Effect size. Also
saves instrument information.
}
